+++ run_tests.py
@@ -2,7 +2,7 @@ import os
 import sys
 import argparse
 from devo.common import load_env_file
-from tests.sender.local_servers import SSL_Server
+from tests.sender.local_servers import SSL_Server, TCP_Server
 
 failed = False
 
@@ -71,9 +71,11 @@ if __name__ == '__main__':
                         default=False, nargs='?', help="Generate coverage.")
     args = parser.parse_args()
     local_ssl_server = SSL_Server()
+    local_tcp_server = TCP_Server()
     if args.coverage:
         CoverageCommand().run()
     else:
         TestCommand().run()
 
     local_ssl_server.close_server()
+    local_tcp_server.close_server()