+++ tests/sender/local_servers.py
@@ -1,8 +1,5 @@
 import threading
 import os
-import ssl
-import socket
-import time
 import socket
 import ssl
 import sys
@@ -88,7 +85,7 @@ class TCP_Server:
             s.listen(5)
             conn, addr = s.accept()
             while not self.shutdown:
-                data = conn.recv(1024).decode()
+                data = conn.recv(5000)
                 conn.send(data)
         else:
             server = SocketServer.TCPServer((ip, port), CustomTCPHandler)