+++ CHANGELOG.md
@@ -5,13 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [x.x.x] - xxxx-xx-xx
+## [1.3.0] - 2018-12-03
 #### Added
  * Use of Token auth and JWT in API, added docs and tests
+ * Local servers for tests
+ * YAML file for Travis CI tests
  
 #### Changed
  * Modify Client init class
-
+ * Fixes in tests of Sender
+ * License file
+ 
+#### Removed
+ * Dockerfile
 
 ## [1.2.1] - 2018-11-26
 #### Added